# phase-2/specs/001-fullstack-todo/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Full-Stack Todo API
  description: JWT-secured REST API for multi-user todo application
  version: 1.0.0
  contact:
    name: Phase-II Todo App

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server (placeholder)

security:
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns API health status (no authentication required)
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy

  /api/{user_id}/tasks:
    parameters:
      - name: user_id
        in: path
        required: true
        description: User ID (must match authenticated user)
        schema:
          type: string
          example: user_abc123

    get:
      summary: List all tasks for user
      description: Returns all tasks owned by the authenticated user
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create new task
      description: Creates a new task for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/{user_id}/tasks/{id}:
    parameters:
      - name: user_id
        in: path
        required: true
        description: User ID (must match authenticated user)
        schema:
          type: string
          example: user_abc123
      - name: id
        in: path
        required: true
        description: Task ID
        schema:
          type: integer
          example: 42

    get:
      summary: Get single task
      description: Returns a specific task owned by the authenticated user
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update task
      description: Updates title and description of a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete task
      description: Permanently deletes a task
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{id}/complete:
    parameters:
      - name: user_id
        in: path
        required: true
        description: User ID (must match authenticated user)
        schema:
          type: string
          example: user_abc123
      - name: id
        in: path
        required: true
        description: Task ID
        schema:
          type: integer
          example: 42

    patch:
      summary: Toggle task completion
      description: Toggles the completed status of a task
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 42
        user_id:
          type: string
          example: user_abc123
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Buy groceries
        description:
          type: string
          nullable: true
          maxLength: 10000
          example: Milk, eggs, bread
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: 2026-02-13T10:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-02-13T10:30:00Z

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Buy groceries
        description:
          type: string
          nullable: true
          maxLength: 10000
          example: Milk, eggs, bread

    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Buy groceries and snacks
        description:
          type: string
          nullable: true
          maxLength: 10000
          example: Milk, eggs, bread, chips

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: Error message

  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Title is required

    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Not authenticated

    Forbidden:
      description: User ID in URL does not match authenticated user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Access forbidden

    NotFound:
      description: Task not found or does not belong to user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Task not found